refactor routes handler
