+++ pkg/websocket/handler.go
@@ -1,0 +1,94 @@
+package websocket
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/http"
+
+	"github.com/go-playground/validator"
+	log "github.com/sirupsen/logrus"
+)
+
+// SetupRoutes handle all routes
+func SetupRoutes(app *App) {
+	log.Trace("Setting up routes")
+
+	http.HandleFunc("/ws", app.WSHandler)
+	http.HandleFunc("/send", app.SendHandler)
+	http.HandleFunc("/healthcheck", app.HealthCheckHandler)
+}
+
+func (a *App) WSHandler(w http.ResponseWriter, r *http.Request) {
+	log.Trace("Serving websocket")
+	conn, err := Upgrade(w, r)
+	if err != nil {
+		log.Error(err)
+		fmt.Fprint(w, "%+V\n", err)
+	}
+
+	client := &Client{
+		Conn: conn,
+	}
+
+	client.Read(a)
+}
+
+var validate = validator.New()
+
+var (
+	ErrorConnectionClosed = errors.New("unable to send: connection closed")
+	ErrorInternalError    = errors.New("unable to send: internal error")
+	ErrorBadRequest       = errors.New("unable to send: bad request")
+)
+
+// SendHandler is used to receive messages from external systems
+func (a *App) SendHandler(w http.ResponseWriter, r *http.Request) {
+	log.Tracef("Receiving message from %q", r.Host)
+	payload := IncomingPayload{}
+	err := json.NewDecoder(r.Body).Decode(&payload)
+	if err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte(ErrorBadRequest.Error()))
+		return
+	}
+
+	err = validate.Struct(payload)
+	if err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte(ErrorBadRequest.Error()))
+		return
+	}
+
+	c, found := a.Pool.Clients[payload.To]
+	if !found {
+		w.WriteHeader(http.StatusForbidden)
+		w.Write([]byte(ErrorConnectionClosed.Error()))
+		return
+	}
+
+	err = c.Send(payload)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Write([]byte(ErrorInternalError.Error()))
+		return
+	}
+
+	w.WriteHeader(http.StatusAccepted)
+}
+
+var (
+	ErrorAWSConnection = errors.New("unable to connect to AWS")
+)
+
+// HealthCheckHandler is used to provide a mechanism to check the service status
+func (a *App) HealthCheckHandler(w http.ResponseWriter, r *http.Request) {
+	err := CheckAWS()
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Write([]byte(ErrorAWSConnection.Error()))
+		return
+	}
+
+	w.WriteHeader(http.StatusOK)
+}
